hw4/assignment_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
# Azure SDKs replacing Google Cloud
azure-identity==1.18.0
azure-servicebus==7.12.3
azure-cosmos==4.7.0
httpx  # Added for calling local endpoints from the consumer